requirements.txt
sqlmodel
bs4
python-dateutil
transformers
selenium
requests
torch
pyfiglet
termcolor
steamlit
questionary
wikidata